LATTERR.ENH: transform lattice_errors into an package and break it into 2 files.
