+++ apsuite/lattice_errors/__init__.py
@@ -1,0 +1,7 @@
+"""."""
+
+from .error_types import DipolesErrors, QuadsErrors, SextsErrors,\
+    QuadsSkewErrors, MultipolesErrors, GirderErrors, BPMErrors
+from .manage_errors import ManageErrors
+
+del error_types, manage_errors
